fix: resolve nc2000 link errors + codec_dev subdir build
- Add HANDYPSP/TAB5_PORT via target_compile_definitions (idf_component.yml cmake_compile_options not processed in ESP-IDF v5.5)
- Guard tab5_stubs.cpp stubs with #ifndef HANDYPSP to avoid multiple-definition with real nc2000 source files
- Fix esp_codec_dev: only build device/es8388, add priv_include INCLUDE_DIRS
- Add -fcommon to root CMakeLists.txt (fix multiple definition)
- flash_capture.py: serial-before-flash burn+log tool
- CHANGELOG.md: record today's fixes

+++ components/nc2000/comm.cpp
@@ -171,7 +171,7 @@ void init_parameters(){
     CYCLES_SECOND = 3686400*(pc1000mode) + 5120*1000*(nc1020mode||nc2000mode)+10240*1000*nc3000mode;
     CYCLES_SECOND *= oc_factor;
     CYCLES_MS = CYCLES_SECOND / 1000;
-    printf("cycles per second is %d\n",CYCLES_SECOND);
+    printf("cycles per second is %lu\n",(unsigned long)CYCLES_SECOND);
     
     // below are not used in new cpu loop,
     // but they are kept for old cpu loop for compare
